beautiful

+++ stdafx.h
@@ -10,7 +10,7 @@ using std::string;
 using std::vector;
 
 typedef vector<vector<glm::vec3>> Patch;
-typedef vector<vector<int>> HeightMapPatch;
+typedef vector<vector<float>> HeightMapPatch;
